clubhouse: Reset actions before calling the callback
The code that deals with the quests' user actions (clicking a button
in a notification's banner) was calling the callback for that action
and reseting the actions afterwards, however, if the mentioned
callback had set up some new actions, they would all be removed as
well by this reset.

This was not noticeable before the code that removes the support for
legacy quests' API (d036d293) because that code also emitted quests'
signals from the idle loop, so the actions' code was delayed a
little bit and thus any new actions would not be removed by the
logic mentioned above.

This patch fixes this issue by resetting the actions before calling
the callback, thus "saving" any newly set up actions from being
reset.

https://phabricator.endlessm.com/T25496