quests: handle game state in BreakingIn and Investigation
These two quests communicate with the level 2 lock from the System app
toolbox. Both quests wait for the state "tried" of the lock, set when
the user tries to open the lock. Unlike other locks, this one doesn't
unlock automatically with the key.

The Investigation quest displays a message, preventing the player to
use the lock.

The BreakingIn quest plays a trapped sequence where Riley falls into a
trap. By setting "trap sequence" state for the lock.

https://phabricator.endlessm.com/T25503