quests: Restore the use of timeout based hints in FirstContact
The First Contact was designed to show hints on a timeout basis, rather
then showing the "Hints" button that every other quest does. With the
port to the non-polling approach, this was modified to use the Hints
button for consistency, but this approach hasn't been tested yet with
users, and thus this patch rolls it back to the original way.

https://phabricator.endlessm.com/T25601