clubhouse: Close + delay quest messages
Messages from quests often just change their text, and replace any
previous messages very quickly; as a result, users may not notice that
the message has changed at all.

While a more sofisticated solution for this is still being developed (in
the Shell), as a temporary measure these changes close any previously
sent messages and delay any new ones a fraction of a second, effectively
making the messages more noticeable than before.

Since this behavior would be annoying for the cases where messages
change as a result to user actions (clicking a button in the
notification), there is no delay in such cases.

https://phabricator.endlessm.com/T25603