libquest: Add show_choices_message Quest method
We have a way to show a message with choices for the user (buttons
in the Shell notification) by using the "choices" option with the
show_message method. However, we need a way to do that and get an
AsyncAction object representing it.
Thus, these changes add a new method "show_choices_message" which
does all that and can be used as, e.g.:

  self.show_choices_message('LAUNCH',
                            ('LEVEL2', print, 'HELLO', 'WORLD'),
                            ('SUCCESS', my_function)).wait()

https://phabricator.endlessm.com/T25622