Add ClubhouseState class
This class is a singleton that's supposed to represent the state of the
Clubhouse to quests. Thus, a new clubhouse_state variable is also added
to the Quest class.

For now, the ClubhouseState only manages the current page for the
Clubhouse, i.e. its "current-page" property is changed whenever the
user changes to a different page in the Clubhouse window.

https://phabricator.endlessm.com/T25578