libquests: Reset Registry before loading a different episode
We need to be able to load a new episode after having previously loaded
a different one. So in order to ensure we don't keep QuestSets from
different episodes at the same time in the Registry this patch keeps
track of which episode is loaded and resets both the list of loaded
QuestSets and the loaded modules when loading a different episode.

https://phabricator.endlessm.com/T25474