clubhouse: Automatically change to a new episode
When a new episode is set in the GlobalStateService, then the Clubhouse
should reflect that. For that reason we reload the episode from the
ClubhousePage when that change happens. The exception to this is when
a quest is running, because in that case we don't want the characters
to eventually disappear from the UI, but instead we check for an
episode change after the quest is finished running.

One could argue that episodes ought to change only when quests are
running anyway, and thus this connection to the signal is not needed,
but the quest(s) that set up the next episode may also do so when they
are initialized (as a sanity check due to e.g. an abortion before the
quest is fully finished); and thus, a new episode can be set up when
the preivous one's quests are being loaded, hence the need for the
connection management.

https://phabricator.endlessm.com/T25474