libquest: Load episodes in the Registry till they're current
Even though a new episode can be set up in the GlobalStateService by a
quest (of a previous episode), it may happen that a quest sets it up
when initialized (for a sanity check, in case the quest was finished
but aborted before registering the new episode). In this case, after
loading an episode, the current episode in the GSS is effectively
not the one loaded, which can be a problem since there's no obvious
way for users to switch to it.

To avoid that, after loading an episode in the Registry, it checks
whether the currently set up episode is still the one just loaded.
Otherwise it will keep loading episodes (preventing a circular setting)
until we find the current one.

https://phabricator.endlessm.com/T25474