libquest: Add the concept of autorun-quest to the Registry
Episodes may have the need to run a quest automatically when they're
loaded, e.g. for the 1st episode to greet the user on their first login
and direct them to the Clubhouse.

Thus, these changes make the Registry look for an AUTORUN_QUEST
variable in the episode's package itself and introduce a new method
get_autorun_quest() to it. This method will be used at a later stage to
look for that quest and run it when appropriate.

https://phabricator.endlessm.com/T25318