clubhouse: Run the Episode's autorun quest automatically
The Registry now contains information about a quest that should be
run automatically when the Clubhouse is launched.

We only run the mentioned quest when the application is run without
any kind of command line options, so we allow the use of its commmands
without starting a quest which is not desired. The exception to this
is the "--debug" option, which should be considered as a "normal" run,
just with turning the debug options on.
The reason for running such a logic like this, instead of just calling
it in the startup implementation, is because we don't want auto-run
quests to be run if users are using CLI options that represent end
functionility on their own (setting an episode, listing episodes,
etc.).

https://phabricator.endlessm.com/T25318