clubhouse: Don't consider unavailable quests as current
When clicking on a character whose quest is running, it will ask the
user whether to stop or continue that quest. This is fine, but for the
Welcome quest case, when it detects that the user has opened the
Clubhouse, it will timeout after a few seconds, and if the user is fast
in clicking the character during that timeout, it'd be better not to
ask that question.

To solve that, this patch skips that question for quests that are
running but not available. This will work for the Welcome quest case,
and since most quests also set themselves as unavailable when they're
complete, it effectively means that characters will not offer the option
to quit when apps have nothing to offer making the UX a bit nicer/more
logical.

https://phabricator.endlessm.com/T25318