Add episode2 teaser poster and new badge
This patch adds the new design for the episode 2 final teaser poster and
a new badge.

The style "completed" is added to the episodes page to be able to set a
different background when the episode is completed and a new background
is added to show the comming soon for the episode 3.

https://phabricator.endlessm.com/T25405