quests: Delay the success message in the Welcome quest (Ep. 2)
The success message in the Welcome quest was being shown right at the
same moment the page change was being shown, and since the Clubhouse
window mayb be sliding in at the same time, the message animation can
seem jerkey. So this patch delays the message a little bit, and this
actually also improves the UX since it's less distracting this way.

https://phabricator.endlessm.com/T25724