quests: Add game state for the visibility of Lightspeed topics
This adds keys to the game state, `lightspeed.topic.*` where `*` can be
`spawnEnemy`, `updateAsteroid`, `updateSpinner`, `updateSquid` or
`updateBeam`. We use the game state service to tell the toolbox to
reveal these topics the first time they are mentioned in each quest, or
at the beginning of the quest if the quest dialog just assumes they are
already revealed.

During normal traversal of the quests, these calls should only do
anything the first time they are encountered. Once topics are revealed,
there is currently nothing that un-reveals them. However, we make sure
to reveal everything we need in every quest where it's relevant. For
example, in case you debug-skip some quests.

The odd one out is the updateAsteroid function, which players will use
during their explorations but is never actually mentioned in any of the
quests. We reveal this one in the LightspeedFix2 quest, simply because
I can't think of a better place to do it, and LightSpeedFix1 already
reveals the spawnEnemy function so I don't want to do both at the same
time.

https://phabricator.endlessm.com/T25677