libquest: Make it easy for Quests to override the "run sound"
The Quest base class was playing the "quest-given" sound when a quest
started running. There may be however a need for this sound to change
per quest, so this patch makes it easy to change it by simply
overriding the value of the Quest's "__sound_on_run_begin__" variable.

This capability will first be used by the Welcome quest to actually
silence itself, since it runs but only displays a message at a later
point, if at all; meaning that without the user hears that the "quest
given" sound, but there's no apparent quest running.

https://phabricator.endlessm.com/T25318