system: Add new set_async method to GameStateService
This new method is a convenient wrapper for the async version of the
GSS' Set method. It will be used in a later commit, and we could
arguably change it to be the default version of "set" but this would
require a bit more testing so let's keep it like this for now.

https://phabricator.endlessm.com/T25759