libquest: Make Quests' save_conf method asynchronous
The save_conf method was being used at the end of every quest and saves
the configuration synchronously, guaranteeing that the GSS set method
gets a result. This call could block a bit, as expected, and while this
is not a problem in most cases, in others it may end up delaying events.
This is the case for the Welcome quest; since it calls save_conf
directly right after the Clubhouse window is set as visible, the method
ended up normally delaying GTK paint events, so users could see the
Clubhouse button sliding left but no apparent window being visible.

After debating whether to offer the two options (sync/async) of
save_conf, it feels like just using the async version is safe for our
use-cases and has the advantage of not blocking ever, so this patch
adds those changes.

https://phabricator.endlessm.com/T25759