clubhouse: Don't run auto-run quests unless on normal sessions
Recently we've introduced the concept of auto-run quets in order to
be able to quests to run automatically when the Clubhouse is started,
and this is being used to show the Welcome quest automatically to the
user.

However, this quest is currently running in any kind of session, and
this is not desired unless on normal user sessions.

We could add this restriction as a setting per app, but for now let's
keep it simple until we figure out a better mechanism to enforce this
and other eventual rules/requirements.

https://phabricator.endlessm.com/T25318