libquest: Ensure the quests' current waiting loop is cancelled
If a quest is currently waiting on an action to be done and it is
cancelled but the action's future object is not resolved/cancelled
itself then that loop is not finished either.

This patch makes sure that a quest's current waiting loop will be
stopped if the quest is cancelled, independently from the state of its
related future objects.

https://phabricator.endlessm.com/T25807