clubhouse: Start running quests only when the loop is idle
Due to how glibcoro works internally, somehow when a quest is
finished and a new one is immediately started, it may end up delaying
events previously scheduled and unrelated this new quest.

A proper fix to this may require more investigation and changes to
glibcoro, so for now we make sure that we only start running a quest
when the mainloop is idle (and thus any pending events have been
dispatched).

https://phabricator.endlessm.com/T25807