libquest: Cancel pause calls when the quest is cancelled
If there's a scheduled call from a pause method, then even if the loop
is stopped, the call does happen after the pause waiting time.

While maybe the cause for the case above should be prevents in
glibcoro, this patch prevents that by directly cancelling the scheduled
pause call handler when the quest is cancelled.

https://phabricator.endlessm.com/T25807