clubhouse: Fix crash when an item is removed from the Inventory
When an item is removed from the inventory, the item's widget is being
attempted to be removed directly from the GTK FlowBox. However, this
widget doesn't keep the items that are added to it directly as children

So in order to fix this, this patch uses the item widget's parent for
the removal.

https://phabricator.endlessm.com/T25875