quests: Fix exception in MakerIntro is the app is closed
The MakerIntro was checking the flingCount property of the Fizzics app
but not giving it a default value. Thus, if the property couldn't be
correctly gathered, a None would be returned and hence compared with an
integer which results in an exception in the Quest.

This patch simply adds the default for that property getter.

https://phabricator.endlessm.com/T25875