clubhouse: Abstract the logic for notifying DBus properties
The logic for notifying the SuggestingOpen and the Visible DBus
properties has a lot in common and should thus be abstracted so it
shares the common parts.

This is done now not only because it's good not to duplicate that logic
but also because we will need to notify another property in a comming
commit.

https://phabricator.endlessm.com/T25690