clubhouse: Add RunningQuest DBus property
The Hack Toolbox app needs to understand what is the currently running
quest in the Clubhouse in order to change its behavior accordingly, and
this may be a requirement for other parts of the system in the future.

Thus, this patch makes that possible by adding a DBus property with the
name of the quest being currently run (or an empty string otherwise).

https://phabricator.endlessm.com/T25690