utils: Do not wrap code messages in mono-space font
For some reason the Shell is not calculating the text size for the
Quest dialogs if they have font sizes that vary, which is the case
when using a mono-space font in code messages, and the result is that
there are ellipsis breaking the text automatically in the wrong place.

To avoid that, this patch removes the use of mono-space fonts and
instead just lets the text be colored.

https://phabricator.endlessm.com/T25800