quests: Instruct users to create Spinner enemies in LightSpeedEnemyA2
The LightSpeedEnemyA2 quest needs Spinner enemies to be created, but
first instructs the user to change the y coordinate of them even when
it detected that no enemies were found.

These changes make it first check for the existence of such enemies,
and if there are none, the quest instructs the user to create some.

https://phabricator.endlessm.com/T25917