quests: Corrupt Hackdex when the Hackdex1 quest becomes available
The Hackdex app was being set as corrupted when the Hackdex1 quest
started, but since the corruption state ends up making the app quit,
this means that the quest advances and then gets aborted after the app
is quit.
This is a problem so this patch just moves the corruption state to when
the quest becomes available (giving the app more time to quit before
the quest is eventually started). This is still not a perfect solution
because the app takes around 8 seconds to be quit, so it's enough time
to launch the Clubhouse and quickly start this quest.

https://phabricator.endlessm.com/T25465