clubhouse: Use a different description for non-key inventory items
When clicking on an inventory item, the item will always display a
description that is related to a key, and the reason was that we used
to only have key items. However, now that we have non-key items
(related to the Steal Device quest, for example), the description when
clicking those should be different. This patch adds a generic but
different description so the inventory doesn't refer to the items as
keys.

https://phabricator.endlessm.com/T25962