quests: Ask to play when in Restart/Continue states in the Enemy A2 quest
The LightSpeedEnemyA2 quest was checking if the game was not in the
flipped state and playing so it checked if the goal was done; this led
to the quest asking the user to change the code again even after the
user flipped back to the game state.

To avoid that, these changes verify whether the game is not in the
flipped state and whether the playing state is set; if it is not, the
quest asks the user to play (and only then it checks the goal).

https://phabricator.endlessm.com/T25963