quests: Account for Continue/Restart states in LightSpeedEnemyA3
If the user has beaten the level or has died, then this quest was not
checking for that state since it's not flipped nor playing, and as a
result would just keep asking the user to change the code.

This patch detects those states and makes sure an appropriate message
is shown.

https://phabricator.endlessm.com/T25963