quests: Intruct user for adding the right enemy in LightSpeedEnemyA3
If the user somehow doesn't have the Spinner enemy configured to be
spawn when the mentioned quest is started, then, upon changing the code
as the quest instructs the user to do, it still says that the enemies
are going beyond the bottom.
This is misguiding, so this patchs makes the quest tell the user that
they need to spawn a Spinner enemy.

https://phabricator.endlessm.com/T25963