quests: Instruct user to create enemy in LightSpeedEnemyC2
If the user starts the LightSpeed app, the mentioned quest assumes that
the Beam enemy is already being used, but that's not the case, since no
enemy is being spawned.

Thus, these changes instruct the user to update the spawn code to have
a Beam enemy before proceeding with the rest of the quest.

https://phabricator.endlessm.com/T25963