clubhouse: Fix displaying an item notification
Due to recent changes in what information is returned by the
QuestItemDB get using its get_item method, there was an error when
trying to assign the values in order to display them in an item popup
message. This patch fixes that in a way that prevents errors if the
number of elements in an item increases again in the future.

https://phabricator.endlessm.com/T25962