quests: Replace common app launch logic by ask_for_app_launch
All the apps are using the same logic for asking for an app to be
launched, giving its icon, and waiting for it to be launched.
That leads to repeated logic, and incoherency in some cases, which
makes it harder to maintain and debug.

These changes use the new ask_for_app_launch method to replace the
mentioned launch logic from all quests.

https://phabricator.endlessm.com/T26076