Support scheduling quests
In a recent requirement from Design, quests should be invite other
quests to the user, so we need to add the ability to "chain" quests.
There are many ways this can be implemented, but since the quests just
want to declare what quest to run next (if any), then this patch
implements the notion of a "scheduled quest".

A scheduled quest is a quest that will be scheduled to run after a
timeout when the current quest is finished running. By default this
timeout is 1 second but with this functionality quests can also e.g.
schedule a quest to run in 1 hour and have that quest ask the user if
they want to do a new quest, play a game, or nothing.

The way libquest offers this functionality is by a new method called
schedule_quest that takes a quest name and the timeout mentioned above,
e.g. self.schedule_quest('AdaQuestSet.LightSpeedTweak', 3) .
Scheduling a quest with an invalid name will cancel any previously
scheduled quest.

https://phabricator.endlessm.com/T25581