libquest: Remove support for the Quests' legacy API
No quests are using the legacy API anymore, and thus we can remove
a lot of code as special handling because of the use of threads in
the legacy code.

This had been done before but was reverted after an indication that
having the legacy API would make it safer for the quest writer until it
was clear that he'd be comfortable with the new API; but it is safe (and
recommendable) to remove it now.

https://phabricator.endlessm.com/T25496