clubhouse: Connect quest closer to where it's run
Recently we changed how quests are run, in order to call the real run
method on them only on idle timeout, so glibcoro has time to dispatch
all its delayed events.
However, the code that connects to the new quest's signals and assigns
its cancellable was still being done before that run call. This causes
problems if we're trying to cancel + run the same quest that is running,
because we're connecting to the quest when it may still be running from
the time before.
This was not an issue since we cannot normally start a quest that is
already running, but it is possible when calling the run-quest
application action, and it will be possible to at a later stage if the
quest being run is actually about to quit.

Thus, to avoid that issue, this patch moves the connection of quests
(and assignation of the cancellable) within the method that calls "run"
on them.

https://phabricator.endlessm.com/T25931