libquest: Add the "stopping" property to the Quest class
The "stopping" property exists to tell the Clubhouse that the Quest is
about to be stopped, and can be used to e.g. prevent the Clubhouse to
offer a quest to be continued if it's about to stop.

The property is by default set to True if the quest's "abort" method has
been called (usually as result of an App quitting when the quest is
relying on it).
It is reset (to False) when the quest is finished running or is started
running (in order to ensure we don't start or keep quests around in a
"stopping" state).

https://phabricator.endlessm.com/T25931