libquest: Stop Quest immediately if it's stopping and in the bg
If a quest has set its new "stopping" property to True (meaning it will
stop soon) and it's set to the background (meaning the user closed its
message popup), then we immediately call the "stop()" method instead of
waiting for whatever future stop call. This ensures that the Clubhouse
doesn't keep quests in the background for longer than necessary.

https://phabricator.endlessm.com/T25931