libquest: Only pause when aborting a Quest if there's a message
If there are is no abort quest message to be shown, then the abort
method shouldn't call a pause, as pausing is just intended to give
the user time to read the message. This guarantees the quest is stopped
as soon as possible if it is aborted and doesn't have an abort message.

https://phabricator.endlessm.com/T25931