clubhouse: Don't offer to continue quests that are stopping
If a quest has declared it is stopping soon (through the new "stopping"
property), then the Clubhouse shouldn't ask users if they want to
continue the quest (when clicking in the quest's character).
Instead, the normal "click" response will occur, as if the quest wasn't
running.

This prevents asking the user to continue a quest that will likely be
stopped in a few seconds or less, which may cause confusion and a bad
UX.

https://phabricator.endlessm.com/T25931