clubhouse: Close the in-window message popup if the window is closed
When users click a character, then usually either a quest is offered to
be started or continued through the in-window message popup.
That popup remained visible even if the window is closed for a while,
and this means that:
 1) It can be confusing to see a quest being offered if the user clicked
    a while ago in the character (and eventually lost the context of the
    intended action);
 2) The message may refer to a context that is no longer valid, e.g.
    offering to continue a quest when that quest has been finished or is
    about to stop.

Thus, to avoid the cases mentioned above, this patch simply closed the
mentioned message popup when the Clubhouse window is itself closed.

https://phabricator.endlessm.com/T25931