libquest: Allow complete property setter/getter to be overriden
Turns out that if we override the setter method like we were doing for
the BreakingIn and FizzicsCode2, it will replace the base class' method
instead!
This is very unfortunate but it's apparently the way that logic works
in PyGObject. Thus, in order to fix the complete property setter/getter
at least, this patch uses private methods for them in the base class
again, but calls the new set_complete/get_complete methods; this means
that subclasses can just redeclare these methods and they get called by
the base class when setting/getting this property.

https://phabricator.endlessm.com/T26083