quests: Do not reset the complete property setter
In the FizzicsCode2 and BreakingIn classes, we were decorating their
"set_complete" methods with the @Quest.complete.setter. This is a way
to override the setter of a base class's property. However, this
instead replaces the setter, so once one of this calls is interpreted,
all the remaining set calls for the "complete" property will be
affected.

This patch drops the mentioned decorator because now just the
declaration of the methods will do the right override.

https://phabricator.endlessm.com/T26083