libquest: Add quest-set property to Quest objects
We managed to get this far without having a way for Quest objects to
refer directly to their QuestSet. This was partially intended because
there's nothing preventing a quest from belonging to more than one
QuestSet, but ATM we only have quests belonging to one single QuestSet.

Thus, since we will be adding a convenience method for quickly getting
the more qualified ID (in the form of QuestSetName.QuestName) from the
quests, let's add the notion of quest-set as a property and set it once
a Quest object is instantiated by the QuestSet.

https://phabricator.endlessm.com/T25581