Support confirmation of scheduled Quests
Quests may want to schedule the next quest but not have to deal with
introducing them for the user, or asking for the user confirmation that
they want to proceed with the scheduled quest. Thus, it's useful if
quests can specify whether the Clubhouse should just use the regular
proposal/confirmation procedure for the scheduled quest.

This patch implements that by:
 * Allowing to specify in the schedule_quest method whether to ask for
   the user's confirmation before running the scheduled Quest;
 * Supporting the confirmation procedure in using the Shell Quest view
   (instead of the in-window message popup).

This gives maximum flexibility and convenience to quests whatever the
way they want to schedule/launch quests.

https://phabricator.endlessm.com/T25581