libquest: Schedule the next quest automatically
Quests have a way of scheduling the next quest to be run, however, the
schedule_quest method does require the quest name to be given. This
makes things very flexible, but essentially what's needed is that if
there's a quest to be offered next, then it will happen without much
effort.

To provide this feature, this patch adds a couple of important changes
in the Quest class:
 * A run_finished method is now called whenever the Quest finishes
   running; by default this method will try to schedule a next quest.
   It is also a very convenient way for Quest objects to call logic
   related to that moment in their execution.
 * A get_next_quest method is now defined and by default first gets the
   next quest available in the quest set, and is used by the
   run_finished method to schedule the next quest. It can be overridden
   by any subclasses if e.g. they always want to suggest the same quest.

https://phabricator.endlessm.com/T25581