clubhouse: Manage quest proposal dismissal
When a quest is being proposed after it's been scheduled, then it
happens in the Shell view. At that moment, the user can still open the
Clubhouse and click in a character, even in the one whose quest is being
offered, and it looks a bit funky.

To avoid that, these changes keep track of whether a quest is being
proposed (using Shell view) and dismiss the view when necessary.

https://phabricator.endlessm.com/T25581