clubhouse: Ensure quest actions maintain their order
The Clubhouse was keeping the actions for Quests (the ones that show up
in the Shell Quest View) in a regular dictionary, but this structure
doesn't keep the items insertion order (prior to 3.6). As a result, the
actions are shown with an order related to their keys, but since the
keys are UUIDs, the order is quite random in the end. While this is
desired in some cases (like the Chore quest's quizz), it is not desired
in other cases and should not happen by default.

To fix this, this patch uses an OrderedDict for storing the actions
instead.

https://phabricator.endlessm.com/T26004