quests: Randomize the Chore quizz answers' position
The Chore's quizz answers used to be shown in a random position every
time the quizz was played, but this was a side-effect from how the quest
choices were stored in the Clubhouse.

That side-effect is not longer happening now that the choices' order is
maintained, so this patch randomizes the quizz answers directly in the
quest, in order to keep the same game play experience as before.

https://phabricator.endlessm.com/T26004