Add new App.connect_props_change method
This method is supposed to connect to whatever properties an App has,
and returns a list of handler IDs. It is created so apps can easily
override it and implement more "custom" property connections.

By default this method just connects to the Clippy object properties (as
App.connect_object_props_change) and this patch also replaces the use
of the older method this one in the Quest's connecting methods.

https://phabricator.endlessm.com/T25974