libquest: Add connect_app_props_changes Quest method
This method is so quests don't have to specify implementation details
when connecting to apps' properties, like when they use
"connect_app_js_props_changes", as this new method can be used to
connect to the default Js object properties, or any other that App
subclasses implement on their own.

It will be used at a later stage, and should ideally replace most uses
of the connect_app_* methods in the future.

https://phabricator.endlessm.com/T25974