apps: Implement the "effectiveLevel" property connection in Fizzics
Even though we already offer an easy way to get the effective level from
the Fizzics App, quests were still connecting to it by using the Js
properties "currentLevel" and "levelSuccess". However, since both of
those properties trigger the resolve the associated AsyncAction in the
quest, when is in the success screen of the level, it would correctly
notify the quets about it and the effective level could be correctly
gathered there; however, when proceeding to the next level, the
"levelSuccess" property is set to false, which also triggers the
notification, and at that point the level may still not have been set
to the next one, so this means that at that moment the quest would get
the level previous to the one it's moving to. To make things more
complicated, it could also happen that the level is thus changed
between the quest getting it (on "levelSuccess as False") and
connecting to the properties again, resulting in the quest not getting
notified of the new level until it's changed again.

To avoid this, we override the "connect_props_changed" in the Fizzics
App, and implement a new connection logic for a custom "effectiveLevel"
property. This guarantees that the quest is only notified when the
"effective level" actually changes, avoiding the race explained above.

https://phabricator.endlessm.com/T25974