quests: Connect FizzicsIntro/Fizzics1 to the effectiveLevel property
These quests were using the "currentLevel" and "levelSuccess" properties
to be notified when the effective level changes, however, as explained in
the previous commit, this leads to a race when the "levelSuccess"
property is reset to False and ends up in the quest not actually
perceiving right level change.

This patch fixes that by connecting to the "effectiveLevel" property that
the Fizzics App now implements.

https://phabricator.endlessm.com/T25974