quests: Don't look for enemies in LightSpeedEnemyA2's wait_for_flip
The LightSpeedEnemyA2 quest was checking the number of spawned enemies
right after the app was launched, and at that moment it can be that no
enemy has been spawned yet, so the quest would assume that no enemies
were in fact created, and guide the users towards creating them.

This patch fixes that by removing the check for enemies from the
wait_for_flip step completely, since the enemies' existence is already
checking in the play step; and instead, the wait_for_flip is just given
a different string message ID to display once the app is flipped.

https://phabricator.endlessm.com/T25917