animation: Support non-looping animations
All of the animations we've had so far had to be looping, however, now
there's the requirement for non-looping animations.

These changes implement that support and allow it to be set up from the
animations' own metadata, in the format:
  "loop": true

By default the loop setting is false, so this doesn't change for
existing animations.

https://phabricator.endlessm.com/T26192