animation: Reset current animations before playing new ones
When playing a new animation, if there was another previously running
then it should be reset (its current frame set to 0). Otherwise, if the
animation is played again, it will start from that frame and this makes
it less predictable for Designers to understand what the UX will be when
playing different animations repeatedly.

This is also especially important for non-looping animations, as without
resetting them, only the last frame would ever be drawn.

https://phabricator.endlessm.com/T26192