animation: Support the concept of anchor point
Until now, AnimationImage widgets were being positioned according to
their top-left corner, as it's a common convention across many UI
toolkits (including GTK).
However, we want to support different size animations for one same
image, and thus, if we need to e.g. center two animations of different
sizes in the same position, the dimensions of the animations will
affect how the placement looks in the end. So we need a way for
"pinning" each animation differently when positioning them, and this
called an "anchor point".

Thus, these changes add this new anchor-point concept to the Animation
and AnimationImage classes, and the value for it can be set up by adding
an entry to the animation's metadata as such:
  "anchor": [150, 100]

The actual use of the anchor for placing the animations will
come in a later commit.

https://phabricator.endlessm.com/T26195