clubhouse: Move QuestSetButton when its position changes
The QuestSetButton's position now is changed according to the anchor
point for the currently loaded animation, but the button itself still
needs to be moved to the new position when new animation (and anchor
point) is loaded.

To accomplish that, these changes turn the QuestSetButton position
from a getter method into a GObject property, and connect to its
notify signal in order to move the button when its position is changed.

https://phabricator.endlessm.com/T26195