clubhouse: Don't reset quest actions for every quest answer
Whenever a button in the Quest Shell view is clicked, it will call an
action in the Clubhouse application and from there derive what is the
associated callback from the quest by checking a key.

After checking whether that key was associated or not, the actions map
was always getting reset; this is a problem because if a button is
clicked very quickly in the Quest Shell view, then the same action is
called twice, so the first time it is called, the actions are reset and
the quest may set up new actions, but the second time the action is
called, it would no longer exist in the actions map, but it'd still
reset it, thus losing any association of keys->callbacks and impeding a
quest from proceeding.

These patch fixes the mentioned issue by ensuring that the actions map
is only reset if the key that triggered the action does exist in it,
otherwise nothing (apart from a debug message) is done.

https://phabricator.endlessm.com/T26043