clubhouse: Implement position concept in Character
Currently the QuestSets declare what position they want for their
characters. The motivation for this was that it would be easier to
control the position and have it customized across episodes if needed.
However, what happens currently is that the QuestSets' base code gets
almost duplicated per episode, and they don't really change the
characters' positions at all, so the result is that whenever we need to
update a character's position, we have to do it as many times as
episodes we have.

A more central and logic place for setting the position of Characters
is together with their animations data, since the animation may
influence the position needed for a character (e.g. if an animation
grows in height, then the position of the character may need to be
compensated).

Thus, this patch loads the position for a Character from a fullbody.json
file from the character's data directory (though the existence of the
file is optional), and offers the position by a get_position method.

The use of the method and Character conf files will be added in a later
commit.

https://phabricator.endlessm.com/T26212