Don't rely on the QuestSet's position for the QuestSetButton
Since we now can have the position for a Character/QuestSetButton
coming from a JSON file in the Character's directory, the position
declared in QuestSets should be made optional.

These changes do that by making the QuestSets' position None by default,
and use the Character's own get_position method if there's no position
declared in the related QuestSet. This means that for now QuestSets can
still override the position for their Character.

https://phabricator.endlessm.com/T26212