Move the page buttons slightly lower in the area
The page selection buttons should be a bit lower vertically, as shown
in the recent mockups of the new charactere positions. This was more
evident now that the "Trapped" character is further lower in the
Clubhouse.

This patch moves the buttons a bit lower as mentioned above, but also
the inventory view, so its margins look visually similar as before
(i.e. same margin above and below the page buttons when in the
inventory view).

https://phabricator.endlessm.com/T26156